\name{survmediation-package}
\alias{survmediation-package}
\docType{package}
\title{
The survmediation package
}
\description{
This is a short summary of the features of \pkg{survmediation}, a package in \R for the analysis of mediation effects based on high-dimensional molecular survival data.
}
\details{
The function \code{\link{metest}} allows to use the mediation analysis approach by Lange & Hansen (2011) for survival data. The implementation uses code presented in the Supplementary Materials of the paper by Lange & Hansen (2011).
}
\author{
Roman Hornung

Maintainer: Roman Hornung
  \email{hornung@ibe.med.uni-muenchen.de}
}
\references{
Lange, T., Hansen, J.V. (2011) Direct and indirect effects in a survival context. Epidemiology 22:575–581.
}
